requirements.txt
paho-mqtt
CoAPthon3
pydantic